requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
gTTS
streamlit-mic-recorder
openai-whisper